Improve arduino board testing (allow specifying boards, libraries and examples to test). Remove all Serial.print lines from Arduino conversions of examples. Can be added manually on Arduinoes that support Serial. Try different sounds on gym interval emulator.

+++ examples/medinfo/common/eve_graph_vertexes.c
@@ -104,7 +104,7 @@ static void graph_draw_line(uint8_t num, uint32_t pos, uint32_t count, uint32_t 
 
             // Draw the second last point if it is not within the filter
             // Then update filter to the second last point
-            if ((abs(((x2 - x1) * (y3 - y2)) + y2 - y1) > filter) || (k == 0))
+            if ((abs((int32_t)(((x2 - x1) * (y3 - y2)) + y2 - y1)) > filter) || (k == 0))
             {
                 y1 = y2;
                 x1 = x2;
